refactor: 비지니스 로직을 가지는 model 패키지 분리
model에 BlackJackGame클래스, player패키지, card패키지로 나누어 저장
