+++ src/test/java/blackjack/GamersTest.java
@@ -3,6 +3,7 @@ package blackjack;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
+import blackjack.model.player.Gamers;
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -27,6 +28,6 @@ public class GamersTest {
 
         gamers.nextGamer();
 
-        assertThat(gamers.getCurrentValue().name).isEqualTo(names.get(1));
+        assertThat(gamers.getCurrentValue().getName()).isEqualTo(names.get(1));
     }
 }
