+++ src/test/java/blackjack/PlayerTest.java
@@ -2,6 +2,7 @@ package blackjack;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
+import blackjack.model.player.Gamer;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
