+++ src/test/java/blackjack/TrumpNumberTest.java
@@ -2,7 +2,7 @@ package blackjack;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import blackjack.trumpcard.TrumpNumber;
+import blackjack.model.card.TrumpNumber;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
